+++ msmexplorer/tests/test_msm_plot.py
@@ -3,7 +3,9 @@ from msmbuilder.msm import MarkovStateModel, BayesianMarkovStateModel
 from matplotlib.axes import SubplotBase
 from seaborn.apionly import JointGrid
 
-from ..plots import plot_pop_resids, plot_msm_network, plot_timescales
+from ..plots import (plot_pop_resids, plot_msm_network, plot_timescales,
+                     plot_implied_timescales)
+from . import PlotTestCase
 
 rs = np.random.RandomState(42)
 data = rs.randint(low=0, high=10, size=100000)
@@ -13,25 +15,36 @@ bmsm = BayesianMarkovStateModel()
 bmsm.fit(data)
 
 
-def test_plot_pop_resids():
-    ax = plot_pop_resids(msm)
+class TestMSMPlot(PlotTestCase):
+    """Test the function(s) that visualize MSMs."""
 
-    assert isinstance(ax, JointGrid)
+    def test_plot_pop_resids(self):
+        ax = plot_pop_resids(msm)
 
+        assert isinstance(ax, JointGrid)
 
-def test_plot_msm_network():
-    ax = plot_msm_network(msm)
+    def test_plot_msm_network(self):
+        ax = plot_msm_network(msm)
 
-    assert isinstance(ax, SubplotBase)
+        assert isinstance(ax, SubplotBase)
 
+    def test_plot_timescales_msm(self):
+        ax = plot_timescales(msm, n_timescales=3, xlabel='x', ylabel='y')
 
-def test_plot_timescales_msm():
-    ax = plot_timescales(msm, n_timescales=3, xlabel='x', ylabel='y')
+        assert isinstance(ax, SubplotBase)
 
-    assert isinstance(ax, SubplotBase)
+    def test_plot_timescales_bmsm(self):
+        ax = plot_timescales(bmsm)
 
+        assert isinstance(ax, SubplotBase)
 
-def test_plot_timescales_bmsm():
-    ax = plot_timescales(bmsm)
-
-    assert isinstance(ax, SubplotBase)
+    def test_plot_implied_timescales(self):
+        lag_times = [1, 50, 100, 250, 500, 1000, 5000]
+        msm_objs = []
+        for lag in lag_times:
+            # Construct MSM
+            msm = MarkovStateModel(lag_time=lag, n_timescales=5)
+            msm.fit(data)
+            msm_objs.append(msm)
+        ax = plot_implied_timescales(msm_objs)
+        assert isinstance(ax, SubplotBase)