+++ msmexplorer/tests/test_tpt_plot.py
@@ -4,6 +4,7 @@ from msmbuilder.msm import MarkovStateModel, BayesianMarkovStateModel
 from matplotlib.axes import SubplotBase
 
 from ..plots import plot_tpaths
+from . import PlotTestCase
 
 rs = np.random.RandomState(42)
 data = rs.randint(low=0, high=10, size=100000)
@@ -14,14 +15,17 @@ bmsm = BayesianMarkovStateModel()
 bmsm.fit(data)
 
 
-def test_plot_tpaths_msm():
-    ax = plot_tpaths(msm, 0, 9)
+class TestTPTPlot(PlotTestCase):
+    """Test the function(s) that visualize TPTs."""
 
-    assert isinstance(ax, SubplotBase)
+    def test_plot_tpaths_msm(self):
+        ax = plot_tpaths(msm, 0, 9)
+
+        assert isinstance(ax, SubplotBase)
 
 
-@SkipTest
-def test_plot_tpaths_bmsm():
-    ax = plot_tpaths(bmsm, 0, 9)
+    @SkipTest
+    def test_plot_tpaths_bmsm(self):
+        ax = plot_tpaths(bmsm, 0, 9)
 
-    assert isinstance(ax, SubplotBase)
+        assert isinstance(ax, SubplotBase)